# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GoBooking Bus Ticket Booking API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingPaymentStatus } from "./bookingPaymentStatus";
import type { BookingStatus } from "./bookingStatus";
import type { Passenger } from "./passenger";
import type { Trip } from "./trip";

export interface Booking {
  id: string;
  bookingRef: string;
  userId: string;
  tripId: string;
  trip: Trip;
  seatIds: string[];
  seatNumbers: string[];
  passengers: Passenger[];
  totalAmount: number;
  paymentMethod: string;
  paymentStatus: BookingPaymentStatus;
  status: BookingStatus;
  contactEmail: string;
  contactPhone: string;
  createdAt: string;
}