# lib/api-zod/src/generated/types/createBookingRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GoBooking Bus Ticket Booking API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBookingRequestPaymentMethod } from "./createBookingRequestPaymentMethod";
import type { Passenger } from "./passenger";

export interface CreateBookingRequest {
  tripId: string;
  seatIds: string[];
  passengers: Passenger[];
  paymentMethod: CreateBookingRequestPaymentMethod;
  contactEmail: string;
  contactPhone: string;
}